+++ core/jvm/src/main/scala/fm/common/XMLUtil.scala
@@ -1,5 +1,7 @@
 /*
- * Copyright 2014 Frugal Mechanic (http://frugalmechanic.com)
+ * Copyright (c) 2019 Frugal Mechanic (http://frugalmechanic.com)
+ * Copyright (c) 2020 the fm-common contributors.
+ * See the project homepage at: https://er1c.github.io/fm-common/
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package fm.common
 
 import com.ctc.wstx.stax.WstxInputFactory
@@ -24,7 +27,7 @@ import org.codehaus.stax2.XMLStreamReader2
 
 object XMLUtil {
   /** Does this look like valid XML (i.e. it starts with an opening XML element)? */
-  def isXML(f: File): Boolean = InputStreamResource.forFile(f).buffered().use{ isXML(_) }
+  def isXML(f: File): Boolean = InputStreamResource.forFile(f).buffered().use { isXML(_) }
 
   /** Does this look like valid XML (i.e. it starts with an opening XML element)? */
   def isXML(is: InputStream): Boolean = isXML(is, true)
@@ -32,15 +35,15 @@ object XMLUtil {
   /** Does this look like valid XML (i.e. it starts with an opening XML element)? */
   def isXML(is: InputStream, useMarkReset: Boolean): Boolean = {
     val markLimit: Int = 1024
-    
+
     if (useMarkReset) {
       require(is.markSupported, "Need an InputStream that supports mark()/reset()")
       is.mark(markLimit)
     }
-    
+
     try {
-      val wrappedIs: InputStream = if (useMarkReset) new BoundedInputStream(is, markLimit) else is
-      withXMLStreamReader2(wrappedIs){ isXML(_) }
+      val wrappedIs: InputStream = if (useMarkReset) new BoundedInputStream(is, markLimit.toLong) else is
+      withXMLStreamReader2(wrappedIs) { isXML(_) }
     } catch {
       case _: Exception => false
     } finally {
@@ -49,28 +52,29 @@ object XMLUtil {
   }
 
   /** Does this look like valid XML (i.e. it starts with an opening XML element)? */
-  def isXML(s: String): Boolean = withXMLStreamReader2(s){ isXML(_) }
+  def isXML(s: String): Boolean = withXMLStreamReader2(s) { isXML(_) }
 
   /** Is this complete and valid XML? */
-  def isValidXML(s: String): Boolean = withXMLStreamReader2(s){ isValidXML(_) }
-  
+  def isValidXML(s: String): Boolean = withXMLStreamReader2(s) { isValidXML(_) }
+
   def detectXMLCharset(is: InputStream): Option[Charset] = detectXMLCharset(is, true)
-  def detectXMLCharset(is: InputStream, useMarkReset: Boolean): Option[Charset] = detectXMLCharsetName(is, useMarkReset).map{ CharsetUtil.forName }
-  
+  def detectXMLCharset(is: InputStream, useMarkReset: Boolean): Option[Charset] =
+    detectXMLCharsetName(is, useMarkReset).map { CharsetUtil.forName }
+
   def detectXMLCharsetName(is: InputStream): Option[String] = detectXMLCharsetName(is, true)
-  
+
   /** If this looks like an XML document attempt to detect it's encoding */
   def detectXMLCharsetName(is: InputStream, useMarkReset: Boolean): Option[String] = {
     val markLimit: Int = 1024
-    
+
     if (useMarkReset) {
       require(is.markSupported, "Need an InputStream that supports mark()/reset()")
       is.mark(markLimit)
     }
-    
+
     try {
-      val wrappedIs: InputStream = if (useMarkReset) new BoundedInputStream(is, markLimit) else is
-      withXMLStreamReader2(wrappedIs){ detectXMLCharsetName(_) }
+      val wrappedIs: InputStream = if (useMarkReset) new BoundedInputStream(is, markLimit.toLong) else is
+      withXMLStreamReader2(wrappedIs) { detectXMLCharsetName(_) }
     } catch {
       case _: Exception => None
     } finally {
@@ -136,9 +140,10 @@ object XMLUtil {
         public static final int NAMESPACE=13;
         public static final int NOTATION_DECLARATION=14;
         public static final int ENTITY_DECLARATION=15;
-        */
+         */
         xmlStreamReader.next() match {
-          case START_ELEMENT | END_ELEMENT | CHARACTERS | CDATA => return false // I think these are the only ones we care about
+          case START_ELEMENT | END_ELEMENT | CHARACTERS | CDATA =>
+            return false // I think these are the only ones we care about
           case _ => // Everything else is okay
         }
       }
@@ -148,7 +153,7 @@ object XMLUtil {
       case _: Exception => false
     }
   }
-  
+
   // Note: This is duplicated in the fm-xml project
   private def inputFactory: WstxInputFactory = {
     val f: WstxInputFactory = new WstxInputFactory()
@@ -168,12 +173,13 @@ object XMLUtil {
     import Resource._
     Resource.using(inputFactory.createXMLStreamReader(is).asInstanceOf[XMLStreamReader2])(f)
   }
-  
-  def rootTag(f: File): String = InputStreamResource.forFile(f).use{ rootTag }
-  
-  private def rootTag(is: InputStream): String = withXMLStreamReader2(is){ xmlStreamReader: XMLStreamReader =>
-    // Skip to the root tag (which is the first START_ELEMENT)
-    while(xmlStreamReader.getEventType != START_ELEMENT) xmlStreamReader.next()
-    xmlStreamReader.getLocalName
-  }
+
+  def rootTag(f: File): String = InputStreamResource.forFile(f).use { rootTag }
+
+  private def rootTag(is: InputStream): String =
+    withXMLStreamReader2(is) { xmlStreamReader: XMLStreamReader =>
+      // Skip to the root tag (which is the first START_ELEMENT)
+      while (xmlStreamReader.getEventType != START_ELEMENT) xmlStreamReader.next()
+      xmlStreamReader.getLocalName
+    }
 }
